Code review changes: Rename _schema_public as _schema_common
