+++ fastavro/_schema_common.py
