+++ fastavro/schema.py
@@ -3,7 +3,7 @@
 from os import path
 import json
 
-from ._schema_public import PRIMITIVES, SCHEMA_DEFS, UnknownType
+from ._schema_common import PRIMITIVES, SCHEMA_DEFS, UnknownType
 
 
 def extract_record_type(schema):
